normal runing gptj class to class

+++ megatron/utils.py
@@ -39,7 +39,7 @@ from collections import deque
 
 def reduce_losses(losses):
     """Reduce a tensor of losses across all GPUs."""
-    reduced_losses = torch.cat([loss.clone().detach().view(1) for loss in losses])
+    reduced_losses = torch.cat([loss.clone().detach().view(1).float() for loss in losses])
     torch.distributed.all_reduce(reduced_losses)
     reduced_losses = reduced_losses / torch.distributed.get_world_size()
     return reduced_losses
